Issue 관련 이벤트 이동
